Consolidate raga tests
