+++ src/js/tests/raga.test.ts
@@ -1,6 +1,6 @@
-import { expect, test } from 'vitest';
+import { expect, test, describe } from 'vitest';
 import { Raga, Pitch } from '../classes';
-import { Raga as ModelRaga } from '@model';
+import { Raga as ModelRaga, Pitch as ModelPitch } from '@model';
 
 const yamanRuleSet = {
   sa: true,
@@ -361,3 +361,280 @@ test('JSON round trip with custom tuning', () => {
   const round = ModelRaga.fromJSON({ ...json, ruleSet: customRuleSet });
   expect(round.toJSON()).toEqual(json);
 })
+
+// Consolidated tests from raga.addition.test.ts
+describe('additional raga utilities', () => {
+  const ruleSet = {
+    sa: true,
+    re: { lowered: true, raised: false },
+    ga: { lowered: false, raised: true },
+    ma: { lowered: true, raised: true },
+    pa: true,
+    dha: { lowered: false, raised: true },
+    ni: { lowered: true, raised: false },
+  };
+
+  const fundamental = 200;
+  const et = (n: number) => 2 ** (n / 12);
+  const expectedRatios = [
+    et(0),
+    et(1),
+    et(4),
+    et(5),
+    et(6),
+    et(7),
+    et(9),
+    et(10),
+  ];
+  const expectedStratified = [
+    et(0),
+    [et(1), et(2)],
+    [et(3), et(4)],
+    [et(5), et(6)],
+    et(7),
+    [et(8), et(9)],
+    [et(10), et(11)],
+  ];
+
+  function computeFreqs(r: ModelRaga, low = 100, high = 800) {
+    const freqs: number[] = [];
+    expectedRatios.forEach((ratio) => {
+      const base = ratio * r.fundamental;
+      const lowExp = Math.ceil(Math.log2(low / base));
+      const highExp = Math.floor(Math.log2(high / base));
+      for (let i = lowExp; i <= highExp; i++) {
+        freqs.push(base * 2 ** i);
+      }
+    });
+    freqs.sort((a, b) => a - b);
+    return freqs;
+  }
+
+  const mapping: Array<[string, string | undefined]> = [
+    ['sa', undefined],
+    ['re', 'lowered'],
+    ['ga', 'raised'],
+    ['ma', 'lowered'],
+    ['ma', 'raised'],
+    ['pa', undefined],
+    ['dha', 'raised'],
+    ['ni', 'lowered'],
+  ];
+
+  test('setRatios and stratifiedRatios with custom rules', () => {
+    const r = new ModelRaga({ ruleSet, fundamental });
+    expect(r.setRatios(ruleSet)).toEqual(expectedRatios);
+    expect(r.ratios.length).toBe(expectedRatios.length);
+    expectedRatios.forEach((ratio, idx) => {
+      expect(r.ratios[idx]).toBeCloseTo(ratio);
+    });
+    expect(r.stratifiedRatios.length).toBe(expectedStratified.length);
+    expectedStratified.forEach((ratio, idx) => {
+      if (Array.isArray(ratio)) {
+        expect(r.stratifiedRatios[idx]).toEqual(ratio);
+      } else {
+        expect(r.stratifiedRatios[idx]).toBeCloseTo(ratio as number);
+      }
+    });
+  });
+
+  test('fromJSON, frequencies and helper mappings', () => {
+    const r = new ModelRaga({ ruleSet, fundamental });
+    const json = r.toJSON();
+    const copy = ModelRaga.fromJSON({ ...json, ruleSet });
+    expect(copy.toJSON()).toEqual(json);
+
+    const freqs = r.getFrequencies();
+    const expectedFreqs = computeFreqs(r);
+    expect(freqs.length).toBe(expectedFreqs.length);
+    freqs.forEach((f, idx) => {
+      expect(f).toBeCloseTo(expectedFreqs[idx]);
+    });
+
+    const chosen = freqs[4];
+    const p = r.pitchFromLogFreq(Math.log2(chosen));
+    expect(p).toBeInstanceOf(ModelPitch);
+    expect(p.frequency).toBeCloseTo(chosen);
+
+    mapping.forEach((tuple, idx) => {
+      expect(r.ratioIdxToTuningTuple(idx)).toEqual(tuple);
+    });
+  });
+
+  test('pitchNumberToScaleNumber edge cases', () => {
+    const r = new ModelRaga({ ruleSet, fundamental });
+    const allowed = r.getPitchNumbers(0, 11);
+    allowed.forEach((pn, idx) => {
+      expect(r.pitchNumberToScaleNumber(pn)).toBe(idx);
+    });
+    const disallowed = [2, 3, 8, 11];
+    disallowed.forEach((pn) => {
+      expect(() => r.pitchNumberToScaleNumber(pn)).toThrow();
+    });
+  });
+
+  test('raga conversion helpers', () => {
+    const r = new ModelRaga({ ruleSet, fundamental });
+    const pitchNumbers = [0, 1, 4, 5, 6, 7, 9, 10];
+    const letters = ['S', 'r', 'G', 'm', 'M', 'P', 'D', 'n'];
+
+    pitchNumbers.forEach((pn, idx) => {
+      expect(r.pitchNumberToSargamLetter(pn)).toBe(letters[idx]);
+      expect(r.pitchNumberToScaleNumber(pn)).toBe(idx);
+      expect(r.scaleNumberToPitchNumber(idx)).toBe(pn);
+      expect(r.scaleNumberToSargamLetter(idx)).toBe(letters[idx]);
+    });
+
+    const len = pitchNumbers.length;
+    expect(r.scaleNumberToPitchNumber(len)).toBe(pitchNumbers[0] + 12);
+    expect(r.scaleNumberToPitchNumber(len + 1)).toBe(pitchNumbers[1] + 12);
+
+    const invalidPNs = [2, 3, 8, 11];
+    invalidPNs.forEach(pn => {
+      expect(r.pitchNumberToSargamLetter(pn)).toBeUndefined();
+      expect(() => r.pitchNumberToScaleNumber(pn)).toThrow();
+    });
+  });
+});
+
+// Consolidated tests from raga.custom.test.ts
+describe('custom rule set utilities', () => {
+  const localRuleSet = {
+    sa: true,
+    re: { lowered: true, raised: false },
+    ga: { lowered: false, raised: true },
+    ma: { lowered: true, raised: true },
+    pa: true,
+    dha: { lowered: true, raised: false },
+    ni: { lowered: false, raised: true },
+  };
+
+  const fundamental = 200;
+  const expectedRatios = [
+    2 ** (0 / 12),
+    2 ** (1 / 12),
+    2 ** (4 / 12),
+    2 ** (5 / 12),
+    2 ** (6 / 12),
+    2 ** (7 / 12),
+    2 ** (8 / 12),
+    2 ** (11 / 12),
+  ];
+
+  const ratioMapping = [
+    ['sa', true, expectedRatios[0]],
+    ['re', false, expectedRatios[1]],
+    ['ga', true, expectedRatios[2]],
+    ['ma', false, expectedRatios[3]],
+    ['ma', true, expectedRatios[4]],
+    ['pa', true, expectedRatios[5]],
+    ['dha', false, expectedRatios[6]],
+    ['ni', true, expectedRatios[7]],
+  ] as const;
+
+  function computeExpectedPitches(r: Raga, low = 100, high = 800) {
+    const pitches: Pitch[] = [];
+    ratioMapping.forEach(([swara, raised, ratio]) => {
+      const freq = ratio * r.fundamental;
+      const lowExp = Math.ceil(Math.log2(low / freq));
+      const highExp = Math.floor(Math.log2(high / freq));
+      for (let i = lowExp; i <= highExp; i++) {
+        pitches.push(
+          new Pitch({
+            swara,
+            oct: i,
+            raised,
+            fundamental: r.fundamental,
+            ratios: r.stratifiedRatios,
+          })
+        );
+      }
+    });
+    pitches.sort((a, b) => a.frequency - b.frequency);
+    return pitches.filter((p) => p.frequency >= low && p.frequency <= high);
+  }
+
+  function computeExpectedFreqs(r: Raga, low = 100, high = 800) {
+    const freqs: number[] = [];
+    expectedRatios.forEach((ratio) => {
+      const base = ratio * r.fundamental;
+      const lowExp = Math.ceil(Math.log2(low / base));
+      const highExp = Math.floor(Math.log2(high / base));
+      for (let i = lowExp; i <= highExp; i++) {
+        freqs.push(base * 2 ** i);
+      }
+    });
+    freqs.sort((a, b) => a - b);
+    return freqs;
+  }
+
+  const pitchNumbersSingleOct = [0, 1, 4, 5, 6, 7, 8, 11];
+
+  test('custom rule set basic functions', () => {
+    const r = new Raga({ ruleSet: localRuleSet, fundamental });
+    expect(r.ruleSetNumPitches).toBe(8);
+    pitchNumbersSingleOct.forEach((pn, idx) => {
+      expect(r.getPitchNumbers(0, 11)[idx]).toBe(pn);
+    });
+    expectedRatios.forEach((ratio, idx) => {
+      expect(r.ratios[idx]).toBeCloseTo(ratio);
+    });
+    const fromSetRatios = r.setRatios(localRuleSet);
+    fromSetRatios.forEach((ratio, idx) => {
+      expect(r.ratios[idx]).toBeCloseTo(ratio);
+    });
+  });
+
+  test('getPitches frequencies and flags', () => {
+    const r = new Raga({ ruleSet: localRuleSet, fundamental });
+    const expected = computeExpectedPitches(r);
+    const result = r.getPitches();
+    expect(result.length).toBe(expected.length);
+    result.forEach((p, idx) => {
+      expect(p.frequency).toBeCloseTo(expected[idx].frequency);
+      expect(p.swara).toEqual(expected[idx].swara);
+      expect(p.oct).toBe(expected[idx].oct);
+      expect(p.raised).toBe(expected[idx].raised);
+    });
+  });
+
+  test('frequency helpers with custom rule set', () => {
+    const r = new Raga({ ruleSet: localRuleSet, fundamental });
+    const expectedFreqs = computeExpectedFreqs(r);
+    const freqs = r.getFrequencies();
+    expect(freqs.length).toBe(expectedFreqs.length);
+    freqs.forEach((f, idx) => {
+      expect(f).toBeCloseTo(expectedFreqs[idx]);
+    });
+
+    const pickFreq = freqs[3];
+    const p = r.pitchFromLogFreq(Math.log2(pickFreq));
+    expect(p.frequency).toBeCloseTo(pickFreq);
+
+    const mapping: Array<[string, string | undefined]> = [
+      ['sa', undefined],
+      ['re', 'lowered'],
+      ['ga', 'raised'],
+      ['ma', 'lowered'],
+      ['ma', 'raised'],
+      ['pa', undefined],
+      ['dha', 'lowered'],
+      ['ni', 'raised'],
+    ];
+    mapping.forEach((tuple, idx) => {
+      expect(r.ratioIdxToTuningTuple(idx)).toEqual(tuple);
+    });
+  });
+});
+
+// Consolidated test from raga.model.strings.test.ts
+test('model raga string getters', () => {
+  const r = new ModelRaga();
+  const pl = r.getPitches({ low: r.fundamental, high: r.fundamental * 1.999 });
+  const solfege = pl.map(p => p.solfegeLetter);
+  const pcs = pl.map(p => p.chroma.toString());
+  const western = pl.map(p => p.westernPitch);
+  expect(r.solfegeStrings).toEqual(solfege);
+  expect(r.pcStrings).toEqual(pcs);
+  expect(r.westernPitchStrings).toEqual(western);
+});
